# Make comparison faster between `SyntaxToken.type` and `SyntaxKind.rawValue`