@jpsim 2015-12-19
Can you please explain this change and why it has a performance impact? I'm perplexed.