@norio-nomura 2015-12-19
Because same `String` instance comparison is faster than different `String` instance comparison.
Comparing pointer is faster than comparing entire memory.